+++ mobile/bookmarklet.js
@@ -1,5 +1,5 @@
 /**
- * PE Tracker - Solutions Mobile (iPhone / iPad iOS) v1.8.1
+ * PE Tracker - Solutions Mobile (iPhone / iPad iOS) v1.8.3
  *
  * Option 1 : Raccourci Apple iOS (Raccourcis / Shortcuts App) - RECOMMANDE SUR CHROME iOS
  * Option 2 : Bookmarklet JavaScript Standard ES5 (Safari iOS)